// src/codex/generated/v2/AccountUpdatedNotification.ts
// GENERATED CODE! DO NOT MODIFY BY HAND!

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AuthMode } from "../AuthMode.js";
import type { PlanType } from "../PlanType.js";

export type AccountUpdatedNotification = { authMode: AuthMode | null; planType: PlanType | null };