// src/codex/generated/v2/CancelLoginAccountResponse.ts
// GENERATED CODE! DO NOT MODIFY BY HAND!

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CancelLoginAccountStatus } from "./CancelLoginAccountStatus.js";

export type CancelLoginAccountResponse = { status: CancelLoginAccountStatus };